Move reducer-related functions into separate file
Avoids cyclic import
